Extract Collapse component into react-ui
