+++ packages/react-ui/src/Collapse/test.js
@@ -1,0 +1,15 @@
+import { render } from "@testing-library/react";
+import React from "react";
+
+import { Collapse } from "./index";
+
+describe("<Collapse />", () => {
+  it("renders", () => {
+    const { container } = render(
+      <Collapse title="Hello">
+        <p>Hi</p>
+      </Collapse>
+    );
+    expect(container).toMatchSnapshot();
+  });
+});
